@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.19.0 <3.0.0"
+  sdk: ^3.5.0
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -27,38 +27,38 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
-  battery_plus: ^5.0.2
-  camera: ^0.10.5+2
+  battery_plus: ^6.0.3
+  camera: ^0.11.0+2
   chewie: ^1.4.0
-  connectivity_plus: ^5.0.2
+  connectivity_plus: ^6.0.5
   cupertino_icons: ^1.0.5
-  device_info_plus: ^9.1.1
+  device_info_plus: ^10.1.2
   file_selector: ^1.0.1
-  fl_chart: ^0.65.0
+  fl_chart: ^0.68.0
   flutter:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
-  geolocator: ^10.1.0
-  grouped_list: ^5.1.2
+  geolocator: ^13.0.1
+  grouped_list: ^6.0.0
   image_picker: ^1.0.5
   infinite_listview: ^1.1.0
   intl: ^0.19.0
-  mobile_scanner: ^3.5.1
-  package_info_plus: ^5.0.1
+  mobile_scanner: ^5.2.1
+  package_info_plus: ^8.0.2
   qr_flutter: ^4.0.0
-  screen_brightness: ^0.2.2
-  sensors_plus: ^4.0.2
-  share_plus: ^7.2.1
+  screen_brightness: ^1.0.1
+  sensors_plus: ^6.0.1
+  share_plus: ^10.0.2
   sticky_headers: ^0.3.0+2
-  url_strategy: ^0.2.0
+  url_strategy: ^0.3.0
   video_player: ^2.6.1
 
 dev_dependencies:
-  flutter_lints: ^3.0.1
+  flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
-  lints: ^3.0.0
+  lints: ^4.0.0
 
 flutter:
 
